% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calDrivingForce.R
\name{calDrivingForce}
\alias{calDrivingForce}
\title{Calculating driving force matrix}
\usage{
calDrivingForce(
  network,
  diffexpgene,
  enrichreactomeRes,
  reactome,
  parallelworker,
  th.path = 0.05,
  th.FC = 3,
  th.enrich.type = "p.adj",
  th.enrich = 0.01,
  th.zscore = 2
)
}
\arguments{
\item{network}{the edge list of signal transduction network.}

\item{diffexpgene}{a list consists with P-value (or P-adjust) and log2 fold change matrix
which indicates whether a gene is differentially expressed.}

\item{enrichreactomeRes}{abnormal pathways for each patient.}

\item{reactome}{Reactome pathway information.}

\item{parallelworker}{number of workers that parallel calculating differentally
expression genes of every patients.}

\item{th.path}{giving threshold of enrichment pathways.}

\item{th.FC}{giving threshold of differentially expression log2 fold change.}

\item{th.enrich.type}{indicating whether P-value or P-adjust used to calculate
enrichment p-value of upstream regulators}

\item{th.enrich}{giving threshold of enrichment p-value.}

\item{th.zscore}{giving threshold of causality z-score.}
}
\value{
A list consists two elements, one of them consists with driving force matrix of every patients,
another consists with individual z-score.
}
\description{
Calculating driving force matrix for each patient
}
\examples{
\dontrun{
drivingforcematrix  <- calDrivingForce(network = STN,
                                      diffexpgene = diffexpgene,
                                      enrichreactomeRes = enrichreactomeRes,
                                      reactome = reactome,
                                      parallelworker = 15)
}
}
\author{
Yan Li
}
