% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPerProcessing.R
\name{dataPerProcessing}
\alias{dataPerProcessing}
\title{Data per-processing for gene expression data and mutation data downloaded from TCGA}
\usage{
dataPerProcessing(
  expressionData,
  mutationData,
  annotationcol,
  removeDuplicates = NULL,
  removeSilenceSNP = F
)
}
\arguments{
\item{expressionData}{a list contains gene expression matrix for NT and TP samples,
which named as 'Normal' and 'Tumor' respactively.}

\item{mutationData}{a dataframe contains mutation terms for each patients.}

\item{annotationcol}{indicate the number of column of gene annotations.}

\item{removeDuplicates}{indicate the column of gene annotation that used to
remove duplicate genes. If there is no need to perform de-duplication, set
removeDuplicates=NULL.}

\item{removeSilenceSNP}{Should remove silence SNPs from mutation data?}
}
\value{
A list with the results after per-processing
}
\description{
Per-processing for raw counts gene expression data and mutation data,
and filter out samples without matched gene expression data and mutation data
}
\examples{
\dontrun{
cancerData <- dataPerProcessing(expressionData = expressionData,
                                mutationData = mutationData,
                                annotationcol = 3,
                                removeDuplicates = 2,
                                removeSilenceSNP = F)
}
}
\author{
Yan Li
}
