% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA.R
\name{DEA}
\alias{DEA}
\title{Personalized differential expression analysis (DEA) based on noise model}
\usage{
DEA(exp, parallelworker, genenamecol = NA, annotationcol = NA)
}
\arguments{
\item{exp}{a list consists with two elements, expression matrix of NT samples
and TP samples. Each elements is a data frame with gene as rows and patient as
colnums.}

\item{parallelworker}{number of workers that parallel calculating differentally
expression genes of every patients.}

\item{genenamecol}{indicate the anotation column that regarded as gene names.}

\item{annotationcol}{indicate the number of column of gene annotations.}
}
\value{
A list consists with the differential expression p-value, p-adj and
log2 fold change.
}
\description{
Differential exprression analysis (DEA) for each patient
}
\examples{
\dontrun{
diffexpgene <- DEA(exp = cancerData$expressionData,
                   parallelworker = 4,
                   genenamecol = 2,
                   annotationcol = 2)
}
}
\author{
Yan Li
}
