% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calDrivingForce.R
\name{getDatafromTCGA}
\alias{getDatafromTCGA}
\title{Download multi-omic data used by PhenoDriver from TCGA}
\usage{
getDatafromTCGA(
  cancerType,
  dataType,
  directory = './GDCdownload/',
  geneType = 'protein_coding',
  cleanDownload = FALSE)
}
\arguments{
\item{cancerType}{select cancer type for which analysis should be run.
\itemize{
\item{ TCGA-ACC }
\item{ TCGA-BLCA }
\item{ TCGA-BRCA }
\item{ TCGA-CESC }
\item{ TCGA-CHOL }
\item{ TCGA-COAD }
\item{ TCGA-DLBC }
\item{ TCGA-ESCA }
\item{ TCGA-GBM }
\item{ TCGA-HNSC }
\item{ TCGA-KICH }
\item{ TCGA-KIRC }
\item{ TCGA-KIRP }
\item{ TCGA-LAML }
\item{ TCGA-LGG }
\item{ TCGA-LIHC }
\item{ TCGA-LUAD }
\item{ TCGA-LUSC }
\item{ TCGA-MESO }
\item{ TCGA-OV }
\item{ TCGA-PAAD }
\item{ TCGA-PCPG }
\item{ TCGA-PRAD }
\item{ TCGA-READ }
\item{ TCGA-SARC }
\item{ TCGA-SKCM }
\item{ TCGA-STAD }
\item{ TCGA-TGCT }
\item{ TCGA-THCA }
\item{ TCGA-THYM }
\item{ TCGA-UCEC }
\item{ TCGA-UCS }
\item{ TCGA-UVM }
}}

\item{dataType}{indicate data type which need to be downloaded. It should be contained one or more of the following item: 'mutation', 'expression', 'clinical', 'all'.}

\item{directory}{Directory/Folder where the data was downloaded. Default: GDCdownload.}

\item{geneType}{Only indicated gene type will be reserved.}

\item{cleanDownload}{Should downloaded data be deleted?}
}
\value{
A list with the results
}
\description{
Download one or more of the following types of data: gene expression data (STAR-Count) and clinical data from TCGA, MC3 mutation data from
}
\examples{
\dontrun{
cancerData <- getDatafromTCGA(cancerType = "BRCA",
                              dataType = 'all',
                              directory = "./data/",
                              cleanDownload = T)
}
}
\author{
Yan Li
}
