% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegulatoryPath.R
\name{getRegulatoryPath}
\alias{getRegulatoryPath}
\title{Get regulatory paths from driver gene to abnormal pathways}
\usage{
getRegulatoryPath(drivergenes, zscore_ind, network, quary, quary_pathway)
}
\arguments{
\item{drivergenes}{A list with individual driver list and ranked population driver list.}

\item{zscore_ind}{A list consist with individual z-score.}

\item{network}{the edge list of signal transduction network.}

\item{quary}{indicating the query driver gene}

\item{quary_pathway}{indicating which abnormal pathways need to be constructed regulatory paths}
}
\value{
edge list of regulatory paths from query driver gene to query abnormal pathways, which can be
used directly as the Cytoscape input to visualize this sub-network.
}
\description{
Reconstructing sub-network which describes the possible regulatory paths from interest driver genes
to its associated abnormal pathways.
}
\examples{
\dontrun{
regulation_path_edgelist <- getRegulatoryPath(drivergenes = drivergenes,
                                              zscore_ind = genescore_individual$zscore,
                                              network = STN,
                                              quary = 'TP53',
                                              quary_pathway = c('EML4 and NUDC in mitotic spindle formation',
                                                                'Resolution of Sister Chromatid Cohesion',
                                                                'Mitotic Prometaphase',
                                                                'RHO GTPases Activate Formins',
                                                                'Separation of Sister Chromatids',
                                                                'Amplification of signal from unattached kinetochores via a MAD2 inhibitory signal'))
}
}
\author{
Yan Li
}
